+++ pkg/api/query/grpc.go
@@ -14,6 +14,7 @@ import (
 	"github.com/prometheus/prometheus/promql"
 	"github.com/prometheus/prometheus/storage"
 
+	"github.com/thanos-io/promql-engine/api"
 	"github.com/thanos-io/promql-engine/engine"
 	"github.com/thanos-io/promql-engine/logicalplan"
 	"github.com/thanos-io/thanos/pkg/api/query/querypb"
@@ -27,7 +28,8 @@ type GRPCAPI struct {
 	now                         func() time.Time
 	replicaLabels               []string
 	queryableCreate             query.QueryableCreator
-	queryFactory                *QueryFactory
+	remoteEndpointsCreate       query.RemoteEndpointsCreator
+	queryCreator                queryCreator
 	defaultEngine               querypb.EngineType
 	lookbackDeltaCreate         func(int64) time.Duration
 	defaultMaxResolutionSeconds time.Duration
@@ -36,8 +38,9 @@ type GRPCAPI struct {
 func NewGRPCAPI(
 	now func() time.Time,
 	replicaLabels []string,
-	creator query.QueryableCreator,
-	queryFactory *QueryFactory,
+	queryableCreator query.QueryableCreator,
+	remoteEndpointsCreator query.RemoteEndpointsCreator,
+	queryCreator queryCreator,
 	defaultEngine querypb.EngineType,
 	lookbackDeltaCreate func(int64) time.Duration,
 	defaultMaxResolutionSeconds time.Duration,
@@ -45,8 +48,9 @@ func NewGRPCAPI(
 	return &GRPCAPI{
 		now:                         now,
 		replicaLabels:               replicaLabels,
-		queryableCreate:             creator,
-		queryFactory:                queryFactory,
+		queryableCreate:             queryableCreator,
+		remoteEndpointsCreate:       remoteEndpointsCreator,
+		queryCreator:                queryCreator,
 		defaultEngine:               defaultEngine,
 		lookbackDeltaCreate:         lookbackDeltaCreate,
 		defaultMaxResolutionSeconds: defaultMaxResolutionSeconds,
@@ -95,10 +99,15 @@ func (g *GRPCAPI) Query(request *querypb.QueryRequest, server querypb.Query_Quer
 		query.NoopSeriesStatsReporter,
 	)
 
+	remoteEndpoints := g.remoteEndpointsCreate(
+		replicaLabels,
+		request.EnablePartialResponse,
+	)
+
 	var qry promql.Query
 	if err := tracing.DoInSpanWithErr(ctx, "instant_query_create", func(ctx context.Context) error {
 		var err error
-		qry, err = g.getInstantQueryForEngine(ctx, request, queryable, maxResolution)
+		qry, err = g.getInstantQueryForEngine(ctx, request, queryable, remoteEndpoints, maxResolution)
 		return err
 	}); err != nil {
 		return err
@@ -106,10 +115,16 @@ func (g *GRPCAPI) Query(request *querypb.QueryRequest, server querypb.Query_Quer
 	defer qry.Close()
 
 	var result *promql.Result
-	tracing.DoInSpan(ctx, "range_query_exec", func(ctx context.Context) {
+	tracing.DoInSpan(ctx, "instant_query_exec", func(ctx context.Context) {
 		result = qry.Exec(ctx)
 	})
 	if result.Err != nil {
+		if request.EnablePartialResponse {
+			if err := server.Send(querypb.NewQueryWarningsResponse(err)); err != nil {
+				return err
+			}
+			return nil
+		}
 		return status.Error(codes.Aborted, result.Err.Error())
 	}
 
@@ -182,10 +197,15 @@ func (g *GRPCAPI) QueryRange(request *querypb.QueryRangeRequest, srv querypb.Que
 		query.NoopSeriesStatsReporter,
 	)
 
+	remoteEndpoints := g.remoteEndpointsCreate(
+		replicaLabels,
+		request.EnablePartialResponse,
+	)
+
 	var qry promql.Query
 	if err := tracing.DoInSpanWithErr(ctx, "range_query_create", func(ctx context.Context) error {
 		var err error
-		qry, err = g.getRangeQueryForEngine(ctx, request, queryable)
+		qry, err = g.getRangeQueryForEngine(ctx, request, queryable, remoteEndpoints, maxResolution)
 		return err
 	}); err != nil {
 		return err
@@ -260,7 +280,13 @@ func extractQueryStats(qry promql.Query) *querypb.QueryStats {
 	return stats
 }
 
-func (g *GRPCAPI) getInstantQueryForEngine(ctx context.Context, request *querypb.QueryRequest, queryable storage.Queryable, maxResolution int64) (promql.Query, error) {
+func (g *GRPCAPI) getInstantQueryForEngine(
+	ctx context.Context,
+	request *querypb.QueryRequest,
+	queryable storage.Queryable,
+	remoteEndpoints api.RemoteEndpoints,
+	maxResolution int64,
+) (promql.Query, error) {
 	lookbackDelta := g.lookbackDeltaCreate(maxResolution * 1000)
 	if request.LookbackDeltaSeconds > 0 {
 		lookbackDelta = time.Duration(request.LookbackDeltaSeconds) * time.Second
@@ -277,8 +303,7 @@ func (g *GRPCAPI) getInstantQueryForEngine(ctx context.Context, request *querypb
 		ts = time.Unix(request.TimeSeconds, 0)
 	}
 	opts := &engine.QueryOpts{
-		LookbackDeltaParam:     lookbackDelta,
-		EnablePartialResponses: request.EnablePartialResponse,
+		LookbackDeltaParam: lookbackDelta,
 	}
 
 	var engineType PromqlEngineType
@@ -297,13 +322,15 @@ func (g *GRPCAPI) getInstantQueryForEngine(ctx context.Context, request *querypb
 	} else {
 		qry = planOrQuery{query: request.Query}
 	}
-	return g.queryFactory.makeInstantQuery(ctx, engineType, queryable, qry, opts, ts)
+	return g.queryCreator.makeInstantQuery(ctx, engineType, queryable, remoteEndpoints, qry, opts, ts)
 }
 
 func (g *GRPCAPI) getRangeQueryForEngine(
 	ctx context.Context,
 	request *querypb.QueryRangeRequest,
 	queryable storage.Queryable,
+	remoteEndpoints api.RemoteEndpoints,
+	maxResolution int64,
 ) (promql.Query, error) {
 	start := time.Unix(request.StartTimeSeconds, 0)
 	end := time.Unix(request.EndTimeSeconds, 0)
@@ -314,17 +341,12 @@ func (g *GRPCAPI) getRangeQueryForEngine(
 		engineParam = g.defaultEngine
 	}
 
-	maxResolution := request.MaxResolutionSeconds
-	if request.MaxResolutionSeconds == 0 {
-		maxResolution = g.defaultMaxResolutionSeconds.Milliseconds() / 1000
-	}
 	lookbackDelta := g.lookbackDeltaCreate(maxResolution * 1000)
 	if request.LookbackDeltaSeconds > 0 {
 		lookbackDelta = time.Duration(request.LookbackDeltaSeconds) * time.Second
 	}
 	opts := &engine.QueryOpts{
-		LookbackDeltaParam:     lookbackDelta,
-		EnablePartialResponses: request.EnablePartialResponse,
+		LookbackDeltaParam: lookbackDelta,
 	}
 
 	var engineType PromqlEngineType
@@ -343,5 +365,5 @@ func (g *GRPCAPI) getRangeQueryForEngine(
 	} else {
 		qry = planOrQuery{query: request.Query}
 	}
-	return g.queryFactory.makeRangeQuery(ctx, engineType, queryable, qry, opts, start, end, step)
+	return g.queryCreator.makeRangeQuery(ctx, engineType, queryable, remoteEndpoints, qry, opts, start, end, step)
 }