+++ pkg/dedup/iter_test.go
@@ -544,7 +544,151 @@ func TestDedupSeriesSet(t *testing.T) {
 			if tcase.isCounter {
 				f = "rate"
 			}
-			dedupSet := NewSeriesSet(&mockedSeriesSet{series: tcase.input}, f)
+			dedupSet := NewSeriesSet(&mockedSeriesSet{series: tcase.input}, f, AlgorithmPenalty)
+			var ats []storage.Series
+			for dedupSet.Next() {
+				ats = append(ats, dedupSet.At())
+			}
+			testutil.Ok(t, dedupSet.Err())
+			testutil.Equals(t, len(tcase.exp), len(ats))
+
+			for i, s := range ats {
+				testutil.Equals(t, tcase.exp[i].lset, s.Labels(), "labels mismatch for series %v", i)
+				res := expandSeries(t, s.Iterator(nil))
+				testutil.Equals(t, tcase.exp[i].samples, res, "values mismatch for series :%v", i)
+			}
+		})
+	}
+}
+
+func TestDedupSeriesSet_Chain(t *testing.T) {
+	for _, tcase := range []struct {
+		name  string
+		input []series
+		exp   []series
+	}{
+		{
+			name: "Single dedup label - exact match",
+			input: []series{
+				{
+					lset:    labels.FromStrings("a", "1", "c", "3"),
+					samples: []sample{{10000, 1}, {20000, 2}, {30000, 3}},
+				},
+				{
+					lset:    labels.FromStrings("a", "1", "c", "3"),
+					samples: []sample{{10000, 1}, {20000, 2}, {30000, 3}},
+				},
+			},
+			exp: []series{
+				{
+					lset:    labels.FromStrings("a", "1", "c", "3"),
+					samples: []sample{{10000, 1}, {20000, 2}, {30000, 3}},
+				},
+			},
+		},
+		{
+			name: "Single dedup label - gap in one series",
+			input: []series{
+				{
+					lset:    labels.FromStrings("a", "1", "c", "3"),
+					samples: []sample{{10000, 1}, {30000, 3}},
+				},
+				{
+					lset:    labels.FromStrings("a", "1", "c", "3"),
+					samples: []sample{{10000, 1}, {20000, 2}, {30000, 3}},
+				},
+			},
+			exp: []series{
+				{
+					lset:    labels.FromStrings("a", "1", "c", "3"),
+					samples: []sample{{10000, 1}, {20000, 2}, {30000, 3}},
+				},
+			},
+		},
+		{
+			name: "Single dedup label - gaps in two series",
+			input: []series{
+				{
+					lset:    labels.FromStrings("a", "1", "c", "3"),
+					samples: []sample{{10000, 1}, {30000, 3}},
+				},
+				{
+					lset:    labels.FromStrings("a", "1", "c", "3"),
+					samples: []sample{{10000, 1}, {20000, 2}},
+				},
+			},
+			exp: []series{
+				{
+					lset:    labels.FromStrings("a", "1", "c", "3"),
+					samples: []sample{{10000, 1}, {20000, 2}, {30000, 3}},
+				},
+			},
+		},
+		{
+			name: "Multi dedup label - exact match",
+			input: []series{
+				{
+					lset:    labels.FromStrings("a", "1", "c", "3"),
+					samples: []sample{{10000, 1}, {20000, 2}, {30000, 3}},
+				},
+				{
+					lset:    labels.FromStrings("a", "1", "c", "3"),
+					samples: []sample{{10000, 1}, {20000, 2}, {30000, 3}},
+				},
+				{
+					lset:    labels.FromStrings("a", "2", "c", "3"),
+					samples: []sample{{10000, 101}, {20000, 102}, {30000, 103}},
+				},
+				{
+					lset:    labels.FromStrings("a", "2", "c", "3"),
+					samples: []sample{{10000, 101}, {20000, 102}, {30000, 103}},
+				},
+			},
+			exp: []series{
+				{
+					lset:    labels.FromStrings("a", "1", "c", "3"),
+					samples: []sample{{10000, 1}, {20000, 2}, {30000, 3}},
+				},
+				{
+					lset:    labels.FromStrings("a", "2", "c", "3"),
+					samples: []sample{{10000, 101}, {20000, 102}, {30000, 103}},
+				},
+			},
+		},
+		{
+			name: "Multi dedup label - gaps in two series",
+			input: []series{
+				{
+					lset:    labels.FromStrings("a", "1", "c", "3"),
+					samples: []sample{{20000, 2}, {30000, 3}},
+				},
+				{
+					lset:    labels.FromStrings("a", "1", "c", "3"),
+					samples: []sample{{10000, 1}, {20000, 2}},
+				},
+				{
+					lset:    labels.FromStrings("a", "2", "c", "3"),
+					samples: []sample{{10000, 101}, {20000, 102}},
+				},
+				{
+					lset:    labels.FromStrings("a", "2", "c", "3"),
+					samples: []sample{{10000, 101}, {30000, 103}},
+				},
+			},
+			exp: []series{
+				{
+					lset:    labels.FromStrings("a", "1", "c", "3"),
+					samples: []sample{{10000, 1}, {20000, 2}, {30000, 3}},
+				},
+				{
+					lset:    labels.FromStrings("a", "2", "c", "3"),
+					samples: []sample{{10000, 101}, {20000, 102}, {30000, 103}},
+				},
+			},
+		},
+	} {
+		t.Run(tcase.name, func(t *testing.T) {
+			dedupSet := NewSeriesSet(&mockedSeriesSet{series: tcase.input}, "", AlgorithmChain)
 			var ats []storage.Series
 			for dedupSet.Next() {
 				ats = append(ats, dedupSet.At())